// src/interfaces/enttiry/human.ts
export interface HumanNameEntity {
  name: string;
  prio: number;
}

export interface HumanAgeEntity {
  age: number;
  prio: number;
}

export interface HumanStatusEntity {
  id: number;
  body: number;
  mental: number;
}

export interface HumanCommandLogEntity {
  id: number;
  score: number;
}